pyproject.toml
[project]
name = "my-first-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "docx2txt>=0.9",
    "faiss-cpu>=1.12.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "pdfminer>=20191125",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.0",
    "tiktoken>=0.11.0",
    "unstructured>=0.18.14",
]
